Structures Spawn Units
Problems to fix: When spawning a unit from any structure it spawns it and if the coroutine stops I cannot spawn anymore.
When starting to spawn a unit at a building for example (Workshop) A catapult is spawned, but if I try spawning at multiple buildings it keeps spawning the catapult as that was the first to be spawned.